[Ref] Simplify Annotator - move DOM-related methods into HtmlDomUtils.ts
